Address deprecation warnings to fix build
There was a CMake error when building from source with Rolling and addressing the deprecation warnings fixed it.
The error was cryptic, so I'm not sure exactly what was going wrong. Here it is for reference:

```
CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
/home/jacob/ws/src/ros-perception/image_transport_plugins/theora_image_transport/include_directories
   used as include directory in directory /home/jacob/ws/src/ros-perception/image_transport_plugins/theora_image_transport
   used as include directory in directory /home/jacob/ws/src/ros-perception/image_transport_plugins/theora_image_transport
```

Signed-off-by: Jacob Perron <jacob@openrobotics.org>